moved stuff around to un-piss off jest. Every cypress-related asset is now in the folder cypress
